Compartments: add portHAS_COMPARTMENT support derived from portHAS_SANDBOX
The implementation of compartments support copies most of the sandboxing
support already with a few changes in design/implementation mainly to allow
compartment calls to nest.

The reason I copy rather than modify the existing sandboxing support is to
avoid any breakage or assumptions that apps using sandboxing have while having
the flexibility to change things.

The main differences (for now) are:
* Upon a compartment entry, the context is emitted in the caller's stack rather
than a special allocated area for a sandbox context.
* Each compartment is associated with a CompID.
* Only sandboxing or compartment features can be used at one time.
* The otype of FreeRTOS return code/data is 4095 (0xFFF) not to conflict with
generated compartments IDs starting from 0.