compartmentalization/cheri: Rework comartment entery trampoline
This trampoline is being used as a template for all functions in a compartmnet.
It acts a first domain switcher function call before actually calling the
intended function. The loader is supposed to allocate this template for each
function, populate it with function cap, captable and the compartment ID.

This is currently unoptimised, but it removes the need to trap and handle
sealed calls as compartment entries.