compartmentalization: Rework compartment-switch trampolines
Split the trampoline to a smaller one that sets up and loads destination compartment context then
jumps the a global shared template trampoline that performas the switch.
The smaller trampoline is only emitted for inter-compartment calls and passed function pointers.
Also remove the same_compartment check as libdl is now able to detect inter-compartment calls.