compartmentalization: Emulate MPU-based compartmentalization
This is just a best-effort emulation assuming unlimited number of MPU regions to be compared with
CHERI-based compartmentalization. The aim is to get comparable and somewhat realistic workloads
and performance overheads by emulating system calls (to reconfigure the MPU), loading MPU contexts
per compartment, and writing to PMP CSRs. The actual protection policy is not enforced on the
the hardware level as the underlying MPU implementation is not scalable/sufficient to protect large
number of regions required by libdl to protect all symbols and sections. Instead zero values are
written to CSR pmpaddrx and pmpcfg registers (which means the regions are disabled/unmatched), and
this should get us CSR writes overheads (i.e., flushing the pipeline or cycles taken to write to
CSRs). The emulation is also all working in M-mode solely because all accesses will fail in U-mode
if there are no matching PMP entries, thus syscalls are emulated from M-mode to M-mode and since
the PMP regions aren't locked, no protection checks will be done in M-mode.